Add files via upload
Project Prerequisites:
IDE Used: NetBeans 11.2 (You can use Eclipse or any other Java editor)
Java should be installed on the machine.
To build a notepad text editor using java we require basic knowledge of java and file operations.
Java provides by default packages such as Abstract Window Toolkit (AWT) & Swing packages to create user interface(UI) for desktop applications.


Import packages
In this step, we will import File, AWT & Swing required packages

import java.io.*;
import javax.swing.*;
import java.awt.*;
import java.awt.event.*;
 


Initialize User Interface
In this very initial step, we will create the basic user interface for our Notepad Desktop Java Application. We will create a menu such as File, Edit, Format & Sub-Menu in that such as New File, Open File, Save File, Exit, Select All Text, Cut , Copy, Paste, Change Font Family, Change Font Style, Change Font Size, etc.

Functions Definitions:

a) setLayout(layout): This function will define the layout of the frame, window, pane, in which all the components are arranged.
b) setText(“your text”): This function will set the title or the text on the label, button,etc.
c) setVisible(true): This function will make the frame or window visible to the user. By default, it is false.
d) setSize(int width, int height): This function takes two parameters such as height and width. It is used to define the size of frame/window, panel, etc.
e) add(obj): This function is used to add the components in a sequential manner to the frame or panel.
 h) setFont(Font): This function sets the font family, font style, font size of the textarea and textfield.
 


Performing Actions of User:
In this step, we will specify what will happen when a particular button is clicked or any shortcut is pressed.

If a user clicks on the New button or presses key Ctrl +N then a new Text File will be created for the user.
 If a user clicks on the Save button or presses Ctrl +S it will allow the user to save the file on their machine.
 If a user clicks on the Copy button or presses Ctrl + C it will allow the user to copy text from the editor.
If a user clicks on the Paste button or presses Ctrl + V it will allow the user to paste text from the java text editor.
If a user clicks on the Cut button or presses Ctrl + X it will allow the user to cut that text from the editor.
If a user clicks on the Select All button or presses Ctrl + A it will allow the user to select all text from the java text editor.
If a user clicks on the Font Family button it will allow the user to format the font family of the text.
If a user clicks on the Font Style button it will allow the user to format the font style of the text.
If a user clicks on the Font Size button it will allow the user to format the font size of the text.
 

Summary

 We have finally built our Notepad – Text Editor in java. Now we can create, edit our text files on our machine: desktop / laptop.

Also, we can now important notes, to do, maintain a diary, we can also write stories, etc. From this section, we have learned how to do file operations, and how to create user interface components such as frame, panels, buttons, labels, filechooser, etc with the help of AWT and Swing package